requirements.txt
langchain
langchain-ollama
python-dotenv
langchain_community
langchain_core
streamlit
bs4
ipykernel
arxiv
pypdf
pymupdf
wikipedia
langchain-text-splitters
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
langchain-groq
duckdb
pandas
langchain-chains